+++ src/DomainLib/AssemblyInfo.fs
@@ -5,8 +5,8 @@ open System.Reflection
 [<assembly: AssemblyTitleAttribute("DomainLib")>]
 [<assembly: AssemblyProductAttribute("DependentTypes")>]
 [<assembly: AssemblyDescriptionAttribute("F# DependentTypes")>]
-[<assembly: AssemblyVersionAttribute("0.0.4")>]
-[<assembly: AssemblyFileVersionAttribute("0.0.4")>]
+[<assembly: AssemblyVersionAttribute("0.0.5")>]
+[<assembly: AssemblyFileVersionAttribute("0.0.5")>]
 [<assembly: AssemblyConfigurationAttribute("Release")>]
 do ()
 
@@ -14,6 +14,6 @@ module internal AssemblyVersionInformation =
     let [<Literal>] AssemblyTitle = "DomainLib"
     let [<Literal>] AssemblyProduct = "DependentTypes"
     let [<Literal>] AssemblyDescription = "F# DependentTypes"
-    let [<Literal>] AssemblyVersion = "0.0.4"
-    let [<Literal>] AssemblyFileVersion = "0.0.4"
+    let [<Literal>] AssemblyVersion = "0.0.5"
+    let [<Literal>] AssemblyFileVersion = "0.0.5"
     let [<Literal>] AssemblyConfiguration = "Release"